I have added the folder movingTriangle containing a program that should show a triangle moving from left to right. It uses the external library SDL2 to create an animation through OpenGL ES 2.0. I have chosen to use SDL2 because I have used it before and because it explicitly states that it works on the Raspberry Pi. Unfortunately, it does not work yet on the Raspberry.
I have also used the glad library to obtain the file glad.cpp and the contents of the folder
movingTriangle/gladInclude. It is used to load the OpenGL ES functions used in the program.
The glad library itself does not have to be installed. The only external library that needs
to be installed is SDL2.

We do not have to use OpenGL ES and SDL2 in our project. I have just made the movingTriangle
program to investigate the possibilities.