I have added the circleProgram folder containing a program that shows complete elastic collisions between circular objects. Initially, there are three circles. More circles can be added by holding down the space bar. The program can be stopped by hitting the escape key.
I have made these kind of programs before, so this one is probably not very interesting. At least
I tried to write it down more clearly by using useful programming techniques like the Cheshire cat
idiom, and it works on the Raspberry pi 3. I have tested the program with a screen resolution of
1280x720.